+++ Common/SF2NetworkID.cs
@@ -2,5 +2,6 @@ namespace SuperfluityTwo.Common;
 
 public enum SF2NetworkID : byte
 {
-    SyncPlayerThaumaturgyCycle_S2C = 0
+    JoiningPlayerRequestingSync = 0,
+    SyncPlayerThaumaturgyCycle = 1,
 }